refactor: created a layout for the verifivation pages
